VST3:  Fixed a potential integer overflow in getHashForTUID()

+++ modules/juce_audio_processors/format_types/juce_VST3PluginFormat.cpp
@@ -73,12 +73,12 @@ static int warnOnFailure (int result) noexcept
 //==============================================================================
 static int getHashForTUID (const TUID& tuid) noexcept
 {
-    int value = 0;
+    uint32 value = 0;
 
     for (int i = 0; i < numElementsInArray (tuid); ++i)
-        value = (value * 31) + tuid[i];
+        value = (value * 31) + (uint32) tuid[i];
 
-    return value;
+    return (int) value;
 }
 
 template <typename ObjectType>
